requirements.txt
attrs==21.4.0
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
et-xmlfile==1.1.0
kapteyn @ https://www.astro.rug.nl/software/kapteyn/kapteyn-3.0.tar.gz
kiwisolver==1.3.2
lttb==0.3.1
matplotlib==3.4.3
mpmath==1.2.1
numpy==1.21.2
ObjectListView2==1.0.1
openpyxl==3.0.9
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
Pypubsub==4.0.3
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
tqdm==4.62.3
wxPython==4.1.1
xlrd==2.0.1
yapf==0.32.0
